requirements.txt
customtkinter
pynput
Pillow
pyperclip
packaging
